+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added option to encode/decode exceptions to rollup-memory-range.lua
+
 ### Fixed
 
 - Fixed HTIF iconsole read from machine state